requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
alembic>=0.7.2
pecan>=0.8.0
pbr>=0.6,!=0.7,<1.0
SQLAlchemy>=0.9.7,<=0.9.99
anyjson>=0.3.3
Babel>=1.3
eventlet>=0.16.1,!=0.17.0
requests>=2.2.0,!=2.4.0
jsonrpclib
netaddr>=0.7.12
python-neutronclient>=2.3.11,<3
WebOb>=1.2.3
six>=1.9.0
oslo.config>=1.9.3,<1.10.0  # Apache-2.0
oslo.context>=0.2.0,<0.3.0                     # Apache-2.0
oslo.db>=1.7.0,<1.8.0  # Apache-2.0
oslo.i18n>=1.5.0,<1.6.0  # Apache-2.0
oslo.log>=1.0.0,<1.1.0  # Apache-2.0
oslo.messaging>=1.8.0,<1.9.0  # Apache-2.0
oslo.middleware>=1.0.0,<1.1.0                  # Apache-2.0
oslo.rootwrap>=1.6.0,<1.7.0  # Apache-2.0
oslo.serialization>=1.4.0,<1.5.0               # Apache-2.0
oslo.utils>=1.4.0,<1.5.0                       # Apache-2.0
PyMySQL>=0.6.2  # MIT License
python-barbicanclient>=3.0.1
python-keystoneclient>=1.1.0
python-novaclient>=2.22.0
posix_ipc
pyOpenSSL>=0.11
WSME>=0.6
pyasn1
pyasn1_modules

jinja2
